--- earth_rover_bed_detection/cfg/bed_detector_params.yaml
@@ -1,2 +1,0 @@
-_threshold_time : 0.5
-_twist_threshold: 0.2
