+++ earth_rover_bed_detection/config/bed_detection.yaml
@@ -1,0 +1,2 @@
+threshold_time : 0.5
+threshold_twist: 0.2
